Fixed lot of issues with AllOf support
