+++ lib/src/code_generators/v3/swagger_models_generator_v3.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 import 'package:swagger_dart_code_generator/src/code_generators/swagger_enums_generator.dart';
 import 'package:swagger_dart_code_generator/src/code_generators/swagger_models_generator.dart';
-import 'package:swagger_dart_code_generator/src/code_generators/v3/swagger_enums_generator_v3.dart';
 import 'package:swagger_dart_code_generator/src/extensions/string_extension.dart';
 import 'package:swagger_dart_code_generator/src/models/generator_options.dart';
 import 'package:collection/collection.dart';
@@ -107,151 +106,6 @@ class SwaggerModelsGeneratorV3 extends SwaggerModelsGenerator {
     });
 
     return generateBase(dartCode, fileName, options, result, false);
-  }
-
-  @override
-  List<String> getAllEnumNames(String swaggerFile) {
-    final results = SwaggerEnumsGenerator.getEnumNamesFromRequests(swaggerFile);
-
-    final swagger = jsonDecode(swaggerFile);
-
-    final components = swagger['components'] as Map<String, dynamic>?;
-
-    final schemas = components == null
-        ? null
-        : components['schemas'] as Map<String, dynamic>?;
-
-    final responses = components == null
-        ? null
-        : components['responses'] as Map<String, dynamic>?;
-
-    final requestBodies = components == null
-        ? null
-        : components['requestBodies'] as Map<String, dynamic>?;
-
-    if (schemas != null) {
-      schemas.forEach((className, map) {
-        final mapMap = map as Map<String, dynamic>;
-        if (mapMap.containsKey('enum')) {
-          results.add(SwaggerModelsGenerator.getValidatedClassName(
-              className.capitalize));
-          return;
-        }
-
-        if (mapMap['type'] == 'array' &&
-            mapMap['items'] != null &&
-            mapMap['items']['enum'] != null) {
-          results.add(SwaggerModelsGenerator.getValidatedClassName(
-              className.capitalize));
-          return;
-        }
-
-        Map<String, dynamic>? properties;
-
-        if (mapMap.containsKey('allOf')) {
-          final allOf = mapMap['allOf'] as List<dynamic>;
-          var propertiesContainer = allOf.firstWhereOrNull(
-                  (e) => (e as Map<String, dynamic>).containsKey('properties'))
-              as Map<String, dynamic>?;
-
-          if (propertiesContainer != null) {
-            properties =
-                propertiesContainer['properties'] as Map<String, dynamic>?;
-          } else {
-            properties = map['properties'] as Map<String, dynamic>?;
-          }
-        } else {
-          properties = map['properties'] as Map<String, dynamic>?;
-        }
-
-        if (properties == null) {
-          return;
-        }
-
-        properties.forEach((propertyName, propertyValue) {
-          var property = propertyValue as Map<String, dynamic>;
-
-          if (property.containsKey('enum') ||
-              (property['items'] != null &&
-                  property['items']['enum'] != null)) {
-            results.add(SwaggerModelsGenerator.getValidatedClassName(
-                SwaggerEnumsGeneratorV3().generateEnumName(
-                    SwaggerModelsGenerator.getValidatedClassName(className),
-                    propertyName)));
-          }
-        });
-      });
-    }
-
-    if (responses != null) {
-      responses.forEach((className, map) {
-        final response = responses[className];
-        final content = response['content'] as Map<String, dynamic>?;
-        final firstContent = content?.entries.firstOrNull?.value;
-        final schema = firstContent == null ? null : firstContent['schema'];
-        if (schema != null &&
-            (schema as Map<String, dynamic>).containsKey('enum')) {
-          results.add(className.capitalize);
-          return;
-        }
-        final properties = schema == null
-            ? null
-            : schema['properties'] as Map<String, dynamic>?;
-
-        if (properties == null) {
-          return;
-        }
-
-        properties.forEach((propertyName, propertyValue) {
-          var property = propertyValue as Map<String, dynamic>;
-
-          if (property.containsKey('enum') ||
-              (property['items'] != null &&
-                  property['items']['enum'] != null)) {
-            results.add(SwaggerEnumsGeneratorV3()
-                .generateEnumName(className, propertyName));
-          }
-        });
-      });
-    }
-
-    if (requestBodies != null) {
-      requestBodies.forEach((className, map) {
-        final response = requestBodies[className];
-        final content = response['content'] as Map<String, dynamic>;
-        final firstContent = content.entries.firstOrNull?.value;
-        final schema = firstContent == null ? null : firstContent['schema'];
-        if (schema != null &&
-            (schema as Map<String, dynamic>).containsKey('enum')) {
-          results.add(className.capitalize);
-          return;
-        }
-        final properties = schema == null
-            ? null
-            : schema['properties'] as Map<String, dynamic>?;
-
-        if (properties == null) {
-          return;
-        }
-
-        properties.forEach((propertyName, propertyValue) {
-          var property = propertyValue as Map<String, dynamic>;
-
-          if (property.containsKey('enum') ||
-              (property['items'] != null &&
-                  property['items']['enum'] != null)) {
-            results.add(SwaggerEnumsGeneratorV3()
-                .generateEnumName(className, propertyName));
-          }
-        });
-      });
-    }
-
-    final resultsWithPrefix = results.map((element) {
-      return 'enums.$element';
-    }).toList();
-
-    return resultsWithPrefix;
   }
 
   @override
